requirements.txt
fastapi
uvicorn
python-jose[cryptography]
passlib[bcrypt]
pydantic
pydantic-settings
sqlalchemy
alembic
asyncpg
psycopg2-binary
redis
python-dotenv
httpx==0.24.1
pytest
pytest-asyncio==0.21.0
